+++ packages/provider/src/tasks/frictionless/frictionlessTasks.ts
@@ -24,7 +24,7 @@ import type {
 	IProviderDatabase,
 	Session,
 } from "@prosopo/types-database";
-import type { Rule } from "@prosopo/user-access-policy";
+import type { AccessPolicy } from "@prosopo/user-access-policy";
 import type { ObjectId } from "mongoose";
 import { v4 as uuidv4 } from "uuid";
 import { checkLangRules } from "../../rules/lang.js";
@@ -88,13 +88,14 @@ export class FrictionlessManager extends CaptchaManager {
 	}
 
 	async scoreIncreaseAccessPolicy(
-		accessRule: Rule | null,
+		accessPolicy: AccessPolicy | undefined,
 		baseBotScore: number,
 		botScore: number,
 		tokenId: FrictionlessTokenId,
 	) {
 		const accessPolicyPenalty =
-			accessRule?.score || this.config.penalties.PENALTY_ACCESS_RULE;
+			accessPolicy?.frictionlessScore ||
+			this.config.penalties.PENALTY_ACCESS_RULE;
 		botScore += accessPolicyPenalty;
 		this.logger.info(() => ({
 			msg: "Address has an image captcha config defined",