+++ demos/client-example-server/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @prosopo/client-example-server
 
+## 2.7.6
+### Patch Changes
+
+- Updated dependencies [64b5bcd]
+  - @prosopo/common@3.0.0
+  - @prosopo/dotenv@3.0.0
+  - @prosopo/types@3.0.0
+  - @prosopo/util@3.0.0
+  - @prosopo/server@2.9.2
+
 ## 2.7.5
 ### Patch Changes