+++ dev/flux/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @prosopo/flux
 
+## 2.6.5
+### Patch Changes
+
+- Updated dependencies [64b5bcd]
+  - @prosopo/common@3.0.0
+  - @prosopo/dotenv@3.0.0
+  - @prosopo/util@3.0.0
+  - @prosopo/config@3.0.0
+
 ## 2.6.4
 ### Patch Changes