+++ packages/common/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @prosopo/common
 
+## 3.0.0
+### Major Changes
+
+- 64b5bcd: Access Controls
+
+### Patch Changes
+
+- Updated dependencies [64b5bcd]
+  - @prosopo/locale@3.0.0
+
 ## 2.7.2
 ### Patch Changes