+++ packages/procaptcha-bundle/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @prosopo/procaptcha-bundle
 
+## 3.0.0
+### Major Changes
+
+- 64b5bcd: Access Controls
+
+### Patch Changes
+
+- Updated dependencies [64b5bcd]
+  - @prosopo/common@3.0.0
+  - @prosopo/dotenv@3.0.0
+  - @prosopo/locale@3.0.0
+  - @prosopo/types@3.0.0
+  - @prosopo/util@3.0.0
+  - @prosopo/procaptcha-frictionless@2.6.19
+  - @prosopo/procaptcha-pow@2.7.5
+  - @prosopo/procaptcha-react@2.6.19
+
 ## 2.7.4
 ### Patch Changes