+++ packages/types-env/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @prosopo/types-env
 
+## 2.7.4
+### Patch Changes
+
+- Updated dependencies [64b5bcd]
+  - @prosopo/types-database@3.0.0
+  - @prosopo/common@3.0.0
+  - @prosopo/types@3.0.0
+
 ## 2.7.3
 ### Patch Changes