+++ packages/types/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @prosopo/types
 
+## 3.0.0
+### Major Changes
+
+- 64b5bcd: Access Controls
+
+### Patch Changes
+
+- Updated dependencies [64b5bcd]
+  - @prosopo/common@3.0.0
+  - @prosopo/locale@3.0.0
+
 ## 2.10.0
 ### Minor Changes